+++ Parallelism/AbstractProcessManager.py
@@ -13,7 +13,7 @@ class AbstractPlayersManager(ABC):
         pass
 
     @abstractmethod
-    def sendCommandToPlayer(self, command: VCommands, guild: Guild, forceCreation: bool = False, context: Union[Context, Interaction] = None):
+    async def sendCommandToPlayer(self, command: VCommands, guild: Guild, forceCreation: bool = False, context: Union[Context, Interaction] = None):
         """If the forceCreation boolean is True, then the context must be provided for the Player to be created"""
         pass
