+++ Parallelism/ProcessPlayer.py
@@ -1,10 +1,11 @@
 import asyncio
-from time import time
+from time import sleep, time
 from urllib.parse import parse_qs, urlparse
+from Music.VulkanInitializer import VulkanInitializer
 from discord import VoiceClient
-from asyncio import AbstractEventLoop
-from threading import RLock, Thread
-from multiprocessing import Lock
+from asyncio import AbstractEventLoop, Semaphore, Queue
+from multiprocessing import Process, RLock, Lock, Queue
+from threading import Thread
 from typing import Callable
 from discord import Guild, FFmpegPCMAudio, VoiceChannel
 from Music.Playlist import Playlist
@@ -28,25 +29,30 @@ class TimeoutClock:
         self.__task.cancel()
 
 
-class PlayerThread(Thread):
-    """Player Thread to control the song playback in the same Process of the Main Process"""
+class ProcessPlayer(Process):
+    """Process that will play songs, receive commands from the main process by a Queue"""
 
-    def __init__(self, bot: VulkanBot, guild: Guild, name: str, voiceChannel: VoiceChannel, playlist: Playlist, lock: Lock, guildID: int, voiceID: int) -> None:
-        Thread.__init__(self, name=name, group=None, target=None, args=(), kwargs={})
+    def __init__(self, name: str, playlist: Playlist, lock: Lock, queueToReceive: Queue,  queueToSend: Queue, guildID: int, voiceID: int) -> None:
+        """
+        Start a new process that will have his own bot instance 
+        Due to pickle serialization, no objects are stored, the values initialization are being made in the run method
+        """
+        Process.__init__(self, name=name, group=None, target=None, args=(), kwargs={})
         # Synchronization objects
         self.__playlist: Playlist = playlist
         self.__playlistLock: Lock = lock
+        self.__queueReceive: Queue = queueToReceive
+        self.__queueSend: Queue = queueToSend
+        self.__semStopPlaying: Semaphore = None
         self.__loop: AbstractEventLoop = None
-        self.__playerLock: RLock = RLock()
         # Discord context ID
         self.__guildID = guildID
         self.__voiceChannelID = voiceID
-        self.__guild: Guild = guild
-        self.__bot: VulkanBot = bot
-        self.__voiceChannel: VoiceChannel = voiceChannel
+        # All information of discord context will be retrieved directly with discord API
+        self.__guild: Guild = None
+        self.__bot: VulkanBot = None
+        self.__voiceChannel: VoiceChannel = None
         self.__voiceClient: VoiceClient = None
-
-        self.__downloader = Downloader()
 
         self.__playing = False
         self.__forceStop = False
@@ -54,23 +60,41 @@ class PlayerThread(Thread):
                                'options': '-vn'}
 
     def run(self) -> None:
-        """This method is called automatically when the Thread starts"""
+        """Method called by process.start(), this will exec the actually _run method in a event loop"""
         try:
-            print(f'Starting Player Thread for Guild {self.name}')
+            print(f'Starting Player Process for Guild {self.name}')
+            self.__playerLock = RLock()
             self.__loop = asyncio.get_event_loop_policy().new_event_loop()
             asyncio.set_event_loop(self.__loop)
-            self.__loop.run_until_complete(self._run())
 
+            self.__downloader = Downloader()
+
+            self.__semStopPlaying = Semaphore(0)
+            self.__loop.run_until_complete(self._run())
         except Exception as e:
             print(f'[Error in Process {self.name}] -> {e}')
 
     async def _run(self) -> None:
+        # Recreate the bot instance and objects using discord API
+        self.__bot = await self.__createBotInstance()
+        self.__guild = self.__bot.get_guild(self.__guildID)
+        self.__voiceChannel = self.__bot.get_channel(self.__voiceChannelID)
         # Connect to voice Channel
         await self.__connectToVoiceChannel()
+
         # Start the timeout function
         self.__timer = TimeoutClock(self.__timeoutHandler, self.__loop)
+        # Thread that will receive commands to be executed in this Process
+        self.__commandsReceiver = Thread(target=self.__commandsReceiver, daemon=True)
+        self.__commandsReceiver.start()
+
         # Start a Task to play songs
         self.__loop.create_task(self.__playPlaylistSongs())
+        # Try to acquire a semaphore, it'll be release when timeout function trigger, we use the Semaphore
+        # from the asyncio lib to not block the event loop
+        await self.__semStopPlaying.acquire()
+        # In this point the process should finalize
+        self.__timer.cancel()
 
     def __verifyIfIsPlaying(self) -> bool:
         if self.__voiceClient is None:
@@ -215,33 +239,42 @@ class PlayerThread(Thread):
 
         self.__loop.create_task(self.__playSong(song), name=f'Song {song.identifier}')
 
-    async def receiveCommand(self, command: VCommands) -> None:
-        type = command.getType()
-        args = command.getArgs()
-        print(f'Player Thread {self.__guild.name} received command {type}')
+    def __commandsReceiver(self) -> None:
+        # Forces the Thread that listen to the commands to await this bot instance
+        # to stablish the connection with discord, may delay when running bots in several servers
+        while True:
+            if self.__guildID is not None:
+                break
+            sleep(0.1)
 
-        try:
-            self.__playerLock.acquire()
-            if type == VCommandsType.PAUSE:
-                self.__pause()
-            elif type == VCommandsType.RESUME:
-                await self.__resume()
-            elif type == VCommandsType.SKIP:
-                await self.__skip()
-            elif type == VCommandsType.PLAY:
-                await self.__playPlaylistSongs()
-            elif type == VCommandsType.PREV:
-                await self.__playPrev(args)
-            elif type == VCommandsType.RESET:
-                await self.__reset()
-            elif type == VCommandsType.STOP:
-                await self.__stop()
-            else:
-                print(f'[ERROR] -> Unknown Command Received: {command}')
-        except Exception as e:
-            print(f'[ERROR IN COMMAND RECEIVER] -> {type} - {e}')
-        finally:
-            self.__playerLock.release()
+        while True:
+            command: VCommands = self.__queueReceive.get()
+            type = command.getType()
+            args = command.getArgs()
+            print(f'Player Process {self.__guild.name} received command {type}')
+
+            try:
+                self.__playerLock.acquire()
+                if type == VCommandsType.PAUSE:
+                    self.__pause()
+                elif type == VCommandsType.RESUME:
+                    asyncio.run_coroutine_threadsafe(self.__resume(), self.__loop)
+                elif type == VCommandsType.SKIP:
+                    asyncio.run_coroutine_threadsafe(self.__skip(), self.__loop)
+                elif type == VCommandsType.PLAY:
+                    asyncio.run_coroutine_threadsafe(self.__playPlaylistSongs(), self.__loop)
+                elif type == VCommandsType.PREV:
+                    asyncio.run_coroutine_threadsafe(self.__playPrev(args), self.__loop)
+                elif type == VCommandsType.RESET:
+                    asyncio.run_coroutine_threadsafe(self.__reset(), self.__loop)
+                elif type == VCommandsType.STOP:
+                    asyncio.run_coroutine_threadsafe(self.__stop(), self.__loop)
+                else:
+                    print(f'[ERROR] -> Unknown Command Received: {command}')
+            except Exception as e:
+                print(f'[ERROR IN COMMAND RECEIVER] -> {type} - {e}')
+            finally:
+                self.__playerLock.release()
 
     def __pause(self) -> None:
         if self.__voiceClient is not None:
@@ -265,12 +298,16 @@ class PlayerThread(Thread):
                     self.__playlist.loop_off()
                     self.__playlist.clear()
 
+                # Send a command to the main process put this to sleep
+                sleepCommand = VCommands(VCommandsType.SLEEPING)
+                self.__queueSend.put(sleepCommand)
                 self.__voiceClient.stop()
                 await self.__voiceClient.disconnect()
 
                 self.__songPlaying = None
                 self.__playing = False
                 self.__voiceClient = None
+                self.__semStopPlaying.release()
             # If the voiceClient is not None we finish things
             else:
                 await self.__forceBotDisconnectAndStop()
@@ -316,8 +353,18 @@ class PlayerThread(Thread):
                 self.__playlist.clear()
                 self.__playlist.loop_off()
 
+    async def __createBotInstance(self) -> VulkanBot:
+        """Load a new bot instance that should not be directly called."""
+        initializer = VulkanInitializer(willListen=False)
+        bot = initializer.getBot()
+
+        await bot.startBotCoro(self.__loop)
+        await self.__ensureDiscordConnection(bot)
+        return bot
+
     async def __timeoutHandler(self) -> None:
         try:
+            # If there is not voiceClient return
             if self.__voiceClient is None:
                 return
 
@@ -336,6 +383,11 @@ class PlayerThread(Thread):
                 with self.__playlistLock:
                     self.__playlist.loop_off()
                 await self.__forceBotDisconnectAndStop()
+                # Send command to main process to finish this one
+                sleepCommand = VCommands(VCommandsType.SLEEPING)
+                self.__queueSend.put(sleepCommand)
+                # Release semaphore to finish process
+                self.__semStopPlaying.release()
         except Exception as e:
             print(f'[ERROR IN TIMEOUT] -> {e}')
 
@@ -348,6 +400,13 @@ class PlayerThread(Thread):
         except Exception as e:
             print(f'[ERROR IN CHECK BOT ALONE] -> {e}')
             return False
+
+    async def __ensureDiscordConnection(self, bot: VulkanBot) -> None:
+        """Await in this point until connection to discord is established"""
+        guild = None
+        while guild is None:
+            guild = bot.get_guild(self.__guildID)
+            await asyncio.sleep(0.2)
 
     async def __connectToVoiceChannel(self) -> bool:
         try:
