+++ Parallelism/ThreadPlayer.py
@@ -1,12 +1,11 @@
 import asyncio
-from time import sleep, time
+from time import time
 from urllib.parse import parse_qs, urlparse
-from Music.VulkanInitializer import VulkanInitializer
-from discord import Member, VoiceClient
-from asyncio import AbstractEventLoop, Semaphore, Queue
-from multiprocessing import Process, RLock, Lock, Queue
-from threading import Thread
-from typing import Callable, List
+from discord import VoiceClient
+from asyncio import AbstractEventLoop
+from threading import RLock, Thread
+from multiprocessing import Lock
+from typing import Callable, Coroutine
 from discord import Guild, FFmpegPCMAudio, VoiceChannel
 from Music.Playlist import Playlist
 from Music.Song import Song
@@ -29,30 +28,26 @@ class TimeoutClock:
         self.__task.cancel()
 
 
-class PlayerProcess(Process):
-    """Process that will play songs, receive commands from the main process by a Queue"""
+class ThreadPlayer(Thread):
+    """Player Thread to control the song playback in the same Process of the Main Process"""
 
-    def __init__(self, name: str, playlist: Playlist, lock: Lock, queueToReceive: Queue,  queueToSend: Queue, guildID: int, voiceID: int) -> None:
-        """
-        Start a new process that will have his own bot instance 
-        Due to pickle serialization, no objects are stored, the values initialization are being made in the run method
-        """
-        Process.__init__(self, name=name, group=None, target=None, args=(), kwargs={})
+    def __init__(self, bot: VulkanBot, guild: Guild, name: str, voiceChannel: VoiceChannel, playlist: Playlist, lock: Lock, guildID: int, voiceID: int, callbackToSendCommand: Callable, exitCB: Callable) -> None:
+        Thread.__init__(self, name=name, group=None, target=None, args=(), kwargs={})
         # Synchronization objects
         self.__playlist: Playlist = playlist
         self.__playlistLock: Lock = lock
-        self.__queueReceive: Queue = queueToReceive
-        self.__queueSend: Queue = queueToSend
-        self.__semStopPlaying: Semaphore = None
         self.__loop: AbstractEventLoop = None
+        self.__playerLock: RLock = RLock()
         # Discord context ID
-        self.__guildID = guildID
         self.__voiceChannelID = voiceID
-        # All information of discord context will be retrieved directly with discord API
-        self.__guild: Guild = None
-        self.__bot: VulkanBot = None
-        self.__voiceChannel: VoiceChannel = None
+        self.__guild: Guild = guild
+        self.__voiceChannel: VoiceChannel = voiceChannel
         self.__voiceClient: VoiceClient = None
+
+        self.__downloader = Downloader()
+        self.__callback = callbackToSendCommand
+        self.__exitCB = exitCB
+        self.__bot = bot
 
         self.__playing = False
         self.__forceStop = False
@@ -60,41 +55,22 @@ class PlayerProcess(Process):
                                'options': '-vn'}
 
     def run(self) -> None:
-        """Method called by process.start(), this will exec the actually _run method in a event loop"""
+        """This method is called automatically when the Thread starts"""
         try:
-            print(f'Starting Player Process for Guild {self.name}')
-            self.__playerLock = RLock()
-            self.__loop = asyncio.get_event_loop_policy().new_event_loop()
-            asyncio.set_event_loop(self.__loop)
-
-            self.__downloader = Downloader()
-
-            self.__semStopPlaying = Semaphore(0)
+            print(f'Starting Player Thread for Guild {self.name}')
+            self.__loop = self.__bot.loop
             self.__loop.run_until_complete(self._run())
+
         except Exception as e:
             print(f'[Error in Process {self.name}] -> {e}')
 
     async def _run(self) -> None:
-        # Recreate the bot instance and objects using discord API
-        self.__bot = await self.__createBotInstance()
-        self.__guild = self.__bot.get_guild(self.__guildID)
-        self.__voiceChannel = self.__bot.get_channel(self.__voiceChannelID)
         # Connect to voice Channel
         await self.__connectToVoiceChannel()
-
         # Start the timeout function
         self.__timer = TimeoutClock(self.__timeoutHandler, self.__loop)
-        # Thread that will receive commands to be executed in this Process
-        self.__commandsReceiver = Thread(target=self.__commandsReceiver, daemon=True)
-        self.__commandsReceiver.start()
-
         # Start a Task to play songs
         self.__loop.create_task(self.__playPlaylistSongs())
-        # Try to acquire a semaphore, it'll be release when timeout function trigger, we use the Semaphore
-        # from the asyncio lib to not block the event loop
-        await self.__semStopPlaying.acquire()
-        # In this point the process should finalize
-        self.__timer.cancel()
 
     def __verifyIfIsPlaying(self) -> bool:
         if self.__voiceClient is None:
@@ -113,6 +89,7 @@ class PlayerProcess(Process):
                     song = self.__playlist.next_song()
 
             if song is not None:
+                print('Criando song')
                 self.__loop.create_task(self.__playSong(song), name=f'Song {song.identifier}')
                 self.__playing = True
 
@@ -155,7 +132,7 @@ class PlayerProcess(Process):
             self.__timer = TimeoutClock(self.__timeoutHandler, self.__loop)
 
             nowPlayingCommand = VCommands(VCommandsType.NOW_PLAYING, song)
-            self.__queueSend.put(nowPlayingCommand)
+            await self.__callback(nowPlayingCommand, self.__guild.id, song)
         except Exception as e:
             print(f'[ERROR IN PLAY SONG FUNCTION] -> {e}, {type(e)}')
             self.__playNext(None)
@@ -179,11 +156,8 @@ class PlayerProcess(Process):
                     self.__playlist.loop_off()
                     self.__songPlaying = None
                     self.__playing = False
-                    # Send a command to the main process put this one to sleep
-                    sleepCommand = VCommands(VCommandsType.SLEEPING)
-                    self.__queueSend.put(sleepCommand)
-                    # Release the semaphore to finish the process
-                    self.__semStopPlaying.release()
+                    # Send a command to the main process to kill this thread
+                    self.__exitCB(self.__guild.id)
 
     def __verifyIfSongAvailable(self, song: Song) -> bool:
         """Verify the song source to see if it's already expired"""
@@ -239,42 +213,33 @@ class PlayerProcess(Process):
 
         self.__loop.create_task(self.__playSong(song), name=f'Song {song.identifier}')
 
-    def __commandsReceiver(self) -> None:
-        # Forces the Thread that listen to the commands to await this bot instance
-        # to stablish the connection with discord, may delay when running bots in several servers
-        while True:
-            if self.__guildID is not None:
-                break
-            sleep(0.1)
+    async def receiveCommand(self, command: VCommands) -> None:
+        type = command.getType()
+        args = command.getArgs()
+        print(f'Player Thread {self.__guild.name} received command {type}')
 
-        while True:
-            command: VCommands = self.__queueReceive.get()
-            type = command.getType()
-            args = command.getArgs()
-            print(f'Player Process {self.__guild.name} received command {type}')
-
-            try:
-                self.__playerLock.acquire()
-                if type == VCommandsType.PAUSE:
-                    self.__pause()
-                elif type == VCommandsType.RESUME:
-                    asyncio.run_coroutine_threadsafe(self.__resume(), self.__loop)
-                elif type == VCommandsType.SKIP:
-                    asyncio.run_coroutine_threadsafe(self.__skip(), self.__loop)
-                elif type == VCommandsType.PLAY:
-                    asyncio.run_coroutine_threadsafe(self.__playPlaylistSongs(), self.__loop)
-                elif type == VCommandsType.PREV:
-                    asyncio.run_coroutine_threadsafe(self.__playPrev(args), self.__loop)
-                elif type == VCommandsType.RESET:
-                    asyncio.run_coroutine_threadsafe(self.__reset(), self.__loop)
-                elif type == VCommandsType.STOP:
-                    asyncio.run_coroutine_threadsafe(self.__stop(), self.__loop)
-                else:
-                    print(f'[ERROR] -> Unknown Command Received: {command}')
-            except Exception as e:
-                print(f'[ERROR IN COMMAND RECEIVER] -> {type} - {e}')
-            finally:
-                self.__playerLock.release()
+        try:
+            self.__playerLock.acquire()
+            if type == VCommandsType.PAUSE:
+                self.__pause()
+            elif type == VCommandsType.RESUME:
+                await self.__resume()
+            elif type == VCommandsType.SKIP:
+                await self.__skip()
+            elif type == VCommandsType.PLAY:
+                await self.__playPlaylistSongs()
+            elif type == VCommandsType.PREV:
+                await self.__playPrev(args)
+            elif type == VCommandsType.RESET:
+                await self.__reset()
+            elif type == VCommandsType.STOP:
+                await self.__stop()
+            else:
+                print(f'[ERROR] -> Unknown Command Received: {command}')
+        except Exception as e:
+            print(f'[ERROR IN COMMAND RECEIVER] -> {type} - {e}')
+        finally:
+            self.__playerLock.release()
 
     def __pause(self) -> None:
         if self.__voiceClient is not None:
@@ -298,16 +263,12 @@ class PlayerProcess(Process):
                     self.__playlist.loop_off()
                     self.__playlist.clear()
 
-                # Send a command to the main process put this to sleep
-                sleepCommand = VCommands(VCommandsType.SLEEPING)
-                self.__queueSend.put(sleepCommand)
                 self.__voiceClient.stop()
                 await self.__voiceClient.disconnect()
 
                 self.__songPlaying = None
                 self.__playing = False
                 self.__voiceClient = None
-                self.__semStopPlaying.release()
             # If the voiceClient is not None we finish things
             else:
                 await self.__forceBotDisconnectAndStop()
@@ -353,18 +314,8 @@ class PlayerProcess(Process):
                 self.__playlist.clear()
                 self.__playlist.loop_off()
 
-    async def __createBotInstance(self) -> VulkanBot:
-        """Load a new bot instance that should not be directly called."""
-        initializer = VulkanInitializer(willListen=False)
-        bot = initializer.getBot()
-
-        await bot.startBotCoro(self.__loop)
-        await self.__ensureDiscordConnection(bot)
-        return bot
-
     async def __timeoutHandler(self) -> None:
         try:
-            # If there is not voiceClient return
             if self.__voiceClient is None:
                 return
 
@@ -383,11 +334,6 @@ class PlayerProcess(Process):
                 with self.__playlistLock:
                     self.__playlist.loop_off()
                 await self.__forceBotDisconnectAndStop()
-                # Send command to main process to finish this one
-                sleepCommand = VCommands(VCommandsType.SLEEPING)
-                self.__queueSend.put(sleepCommand)
-                # Release semaphore to finish process
-                self.__semStopPlaying.release()
         except Exception as e:
             print(f'[ERROR IN TIMEOUT] -> {e}')
 
@@ -400,13 +346,6 @@ class PlayerProcess(Process):
         except Exception as e:
             print(f'[ERROR IN CHECK BOT ALONE] -> {e}')
             return False
-
-    async def __ensureDiscordConnection(self, bot: VulkanBot) -> None:
-        """Await in this point until connection to discord is established"""
-        guild = None
-        while guild is None:
-            guild = bot.get_guild(self.__guildID)
-            await asyncio.sleep(0.2)
 
     async def __connectToVoiceChannel(self) -> bool:
         try:
@@ -421,9 +360,3 @@ class PlayerProcess(Process):
         except Exception as e:
             print(f'[ERROR CONNECTING TO VC] -> {e}')
             return False
-
-    def __getBotMember(self) -> Member:
-        guild_members: List[Member] = self.__guild.members
-        for member in guild_members:
-            if member.id == self.__bot.user.id:
-                return member
